@@ -1,4 +1,4 @@
-package cn.com.smallcake_utils;
+package com.smallcake.utils;
 
 import android.content.Context;
 import android.content.Intent;
@@ -7,7 +7,8 @@ import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.provider.Settings;
-import android.support.v4.content.FileProvider;
+
+import androidx.core.content.FileProvider;
 
 import java.io.File;
 import java.lang.reflect.Field;
